requirements-free-tier.txt
# Minimal requirements for free tier deployment
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pypdf==3.17.4
python-docx==0.8.11
requests==2.31.0